Rename ProxyRequestCountAdvice and ProxyConnectionCountAdvice (#23450)
* Rename ProxyRequestCountAdvice

* Rename ProxyRequestCountAdvice and ProxyConnectionCountAdvice
