+++ agent/plugins/metrics/core/src/test/java/org/apache/shardingsphere/agent/plugin/metrics/core/advice/ProxyConnectionCountAdviceTest.java
@@ -28,20 +28,25 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
-public final class FrontendChannelInactiveAdviceTest extends MetricsAdviceBaseTest {
+public final class ProxyConnectionCountAdviceTest extends MetricsAdviceBaseTest {
     
-    private final FrontendChannelActiveAdvice activeAdvice = new FrontendChannelActiveAdvice();
-    
-    private final FrontendChannelInactiveAdvice inactiveAdvice = new FrontendChannelInactiveAdvice();
+    private final ProxyConnectionCountAdvice advice = new ProxyConnectionCountAdvice();
     
     @Test
     public void assertBeforeMethod() {
         MockTargetAdviceObject targetObject = new MockTargetAdviceObject();
-        activeAdvice.beforeMethod(targetObject, mock(Method.class), new Object[]{});
-        activeAdvice.beforeMethod(targetObject, mock(Method.class), new Object[]{});
-        inactiveAdvice.beforeMethod(targetObject, mock(Method.class), new Object[]{});
+        advice.beforeMethod(targetObject, mockMethod("channelActive"), new Object[]{});
+        advice.beforeMethod(targetObject, mockMethod("channelActive"), new Object[]{});
+        advice.beforeMethod(targetObject, mockMethod("channelInactive"), new Object[]{});
         assertTrue(MetricsPool.get(MetricIds.PROXY_COLLECTION).isPresent());
         assertThat(((FixtureWrapper) MetricsPool.get(MetricIds.PROXY_COLLECTION).get()).getFixtureValue(), is(1d));
+    }
+    
+    private Method mockMethod(final String methodName) {
+        Method result = mock(Method.class);
+        when(result.getName()).thenReturn(methodName);
+        return result;
     }
 }
