+++ agent/plugins/metrics/type/prometheus/src/main/resources/META-INF/conf/prometheus-proxy-advisors.yaml
@@ -24,19 +24,16 @@ advisors:
       - name: processException
         type: method
   - target: org.apache.shardingsphere.proxy.frontend.netty.FrontendChannelInboundHandler
-    advice: org.apache.shardingsphere.agent.plugin.metrics.core.advice.FrontendChannelActiveAdvice
+    advice: org.apache.shardingsphere.agent.plugin.metrics.core.advice.ProxyConnectionCountAdvice
     pointcuts:
       - name: channelActive
         type: method
-  - target: org.apache.shardingsphere.proxy.frontend.netty.FrontendChannelInboundHandler
-    advice: org.apache.shardingsphere.agent.plugin.metrics.core.advice.FrontendChannelReadAdvice
-    pointcuts:
-      - name: channelRead
+      - name: channelInactive
         type: method
   - target: org.apache.shardingsphere.proxy.frontend.netty.FrontendChannelInboundHandler
-    advice: org.apache.shardingsphere.agent.plugin.metrics.core.advice.FrontendChannelInactiveAdvice
+    advice: org.apache.shardingsphere.agent.plugin.metrics.core.advice.ProxyRequestCountAdvice
     pointcuts:
-      - name: channelInactive
+      - name: channelRead
         type: method
   - target: org.apache.shardingsphere.infra.route.engine.SQLRouteEngine
     advice: org.apache.shardingsphere.agent.plugin.metrics.core.advice.SQLRouteEngineAdvice
